logger/pyproject.toml
[project]
name = "logger"
version = "0.1.0"
description = "An environment logger script"
readme = "README.md"
authors = [
  {name = "Tero Paloheimo", email = "tero.paloheimo@iki.fi"}
]
requires-python = ">=3.13"
dependencies = [
    "pyserial>=3.5",
    "pytz>=2024.2",
    "requests>=2.32.3",
    "ruuvitag-sensor",
]

[tool.ruff]
lint.select = ["B", "D", "E", "F", "G", "I", "N", "PERF", "PIE", "PL", "PTH", "SIM", "RUF", "S", "TRY", "W"]
lint.ignore = ["D203", "D213"]
target-version = "py313"

[tool.uv.sources]
ruuvitag-sensor = { git = "https://github.com/ttu/ruuvitag-sensor.git" }
